requirements.txt
fastapi
python-dateutil
python-dotenv
python-jose
python-multipart
requests
sqlalchemy
psycopg2
uvicorn